Modified the ptypes library to avoid capturing nested exceptions since Py3 does it for us.
This change removes the nested exception logic entirely which means that this
breaks the backtrace for Py2 and related. We're "practically" in the future now
and despite all these years of zero-effort and passive whining that I've put in
to ensure Py2 is still around, Py3 has won. Maybe this codebase will get
refactored to cull out some of this ancient code...then again, maybe not.