Modified the ptypes library so that both generic.alloc and pbinary.type.alloc fall back to instance.set if given an unnamed parameter.
Essentially, the .set method will still always change the value of an
instance and the .alloc method will always be used to reallocate a new
size for an instance. Since reallocating a new size doesn't typically
need an unnamed parameter, we hijack the unnamed parameter farther up in
the class hierarchy so that if we receive an unnamed parameter in a
derived class that should've already processed the unnamed parameters,
we take those parameters and send them directly to the .set method.

The structure and array implementations were also updated to use the
.alloc method instead of calling .set directly when given an immediate
value to assign. By doing it this way, this allows implementations to
add their own .alloc method if they want to allow the caller to change
the size of their instance in some arbitrary way depending on the data
that is received.