Removed what appeared to be some cruft from the parray.terminated class of the ptypes library.
There's really no legitimate reason that any ptype.generic.load method
should be explicitly checking which provider the type is backed by. The
code that was removed was explicitly checking for the "empty" provider
and then setting the array length to 0. This interfered with all classes
that inherited from parray.terminated, specifically when trying to
allocate the parray.block class.

I'm willing to bet that this was hacked-in to avoid an infinite loop
when a non-null-initialized terminator is being checked. I'm not sure
what the right way to fix this would be.. but, since I'm the only user
of this crap, I can just figure that shit out some other day.