Refactored the allocator from the ptypes.pstruct library to append fields instead of loading and replacing them.
Previously, an instance of pstruct.type would be allocated by calling
its super class to load from the "empty" provider. After the class has
had all of its fields loaded from the provider, the allocator would then
proceed to update the fields individually depending on what new values
were being passed to it. When the field definitions for a pstruct.type
are being loaded, each definition can explicitly assign arbitrary
attributes/properties depending on values found in other fields that
have been loaded.

However, when allocating, the caller is allowed to explicitly specify
the type, definition, instance, or value to use for whatever fields they
so choose. This allows one to resize or initialize the value of a chosen
field with the added ability of being able to influence another field
that the chosen one may actually depend on.

When allocating a field set to an explicit value, the implementation
will first perform a regular allocation of the field, and then use
ptypes.generic.set to assign the value to the newly allocated field.
This logic when combined with the allocation process for pstruct.type
can actually result in an explicitly allocated field gaining the
properties determined during the load process, instead of the explicit
properties as determined by the chosen field.

As an example of this, if a field definition inherits its blocksize from
another field and the caller allocates that field with an explicit
value, the field containing the blocksize will be initialized with its
value as loaded by the empty provider, and the field that depends on its
value for its blocksize will be loaded by the empty provider, but then
have its blocksize set to 0. Afterwards, the allocator will then
ptypes.generic.set the explicit value to the new field resulting in the
allocated type never getting its blocksize from any of the fields that
have been explicitly used for allocating.

This was fixed by avoiding the initial load process and instead opting
to construct each field as it becomes needed. This changes the way that
pstruct.type does it allocation. Before, the entire pstruct.type
container was populated with uninitialized fields before the allocated
fields were swapped into their place. Now, fields are instantiated and
allocated as they are needed and in the order of their definition.

If this change breaks anything, the pstruct.type.alloc method can be
tweaked to instantiate each object in the first pass, and then followed
up with the allocation for each instantiated object in the instance.