Modified the osi template library to refactor away its dependence on the __base__ module for each individual layer.
This is essentially a preliminary reimplementation of the entire osi
template library. It splits the upper level protocols from the "network"
module into a distinctly separate module named "transport". The original
implementation used a "__base__" module for each layer which has been
completely removed by this commit. Instead, the definitions that are
used to look up the protocol for each layer has been consolidated within
the "__init__" module for each named layer.

The refactor defines two main classes that each layer can derive from in
order to describe how to lookup the next layer for a protocol. These are
the "layer" class, which is a `ptype.definition`, and the "stackable"
class which is just an abstract class that requires one to implement the
"stackable.layer" method so that it returns a tuple describing how to
lookup the layer that should follow. The tuple is composed of the layer
definition, an identifier (for the definition) or type that is used for
the following layer, and an integer describing the number of remaining
bytes required to complete the entire packet.

This allows for better flexibility due to allowing an implementer to
explicitly choose which osi layer to add a protocol definition to, and
by exposing access to any of the layer-defined protocols when composing
a packet. At the moment, this is only useful for decomposing a packet,
and further research will be required for packet composition.

This commit is preliminary and still contains debugging "prints", as
well as other things to better facilitate its development. Some of the
layers still contain the helper-methods that were used by the previous
implementation. Thus, future commits are expected to remove these
useless methods and clean up the logging once the final interface has
been decided upon.