Fixed some of the definitions in the ndk.verifier library module to allow support for more than one provider dll.
According to the article written by Pavel Yosifovich at
https://scorpiosoftware.net/2024/06/01/building-a-verifier-dll/, the
application verifier dll supports more than one provider and determines
the termination of the provider array by checking if the last element is
all nulls. It's obvious that I got lucky when this was first committed.

Anyways, this updates the array type so that it supports multiple
providers and adds an allocator that sets the right size for the
structure that you're supposed to write to DllMain's lpReserved.