Added support for abstract base classes to the `ptypes.providers` library module.
This modifies the base classes for the available providers to use
abstract base classes in Py2 and Py3. This way the user gets a complaint
if a method has the wrong name or is unimplemented, and is forced to
correct their method names in response to it.

This commit supercedes the prior commit that renames the `remote.read`
method to `remote.receive`.