Updated the ptypes.pbinary.container types so that they allocate similarly to the ptypes.ptype.container types.
When allocating an instance of the `pstruct.type` type, a dictionary can
be used which will result in the type being allocated as empty with the
parameter being passed to the `pstruct.type.set` method. This allows one
to pass a list of dictionaries to the `parray.type.alloc` method which
will result in allocating an array of the length specified by the
parameter.

As a result of regular usage, it was discovered that the corresponding
types from the `ptypes.pbinary` library module acted differently. This
commit modifies the `ptypes.pbinary.__structure_interface__` class so
that its "alloc()" method will pass all wildcard and keyword arguments
to its super-class so that the super-class can be responsible for
passing the arguments to the type's "set()" implementation.

The `ptypes.pbinary.array` class was also updated to automatically
allocate with a length that matches the length of the parameter that it
was passed. This results in both `pbinary.container` types matching the
semantics of the regular `ptype.container` types.

Unit tests were added for all additions.