Fixed the setter and allocator for the `pbinary.flags` class from the ptypes library.
The `pbinary.flags` class is essentially the same as `pbinary.struct`
with the exception that single-bit fields can be assigned a boolean. As
a result, the default implementation of both the "set()" and "alloc()"
methods allows you to give them an iterable as the parameter which will
then be used to set or clear the specified flags.

Due to an oversight, when either of these methods are given a dictionary
the original bit-setting logic will be incorrectly used. This results in
the integers in said dictionary being ignored since dictionaries are
also considered iterable.

This commit adds explicit support for dictionaries to both the "set()"
and "alloc()" methods for the `pbinary.flags` class. This way it can be
treated as a regular `pbinary.struct` if necessary. The implementation
will only attempt to set or clear the entire field iff the value for the
field is of a "bool" type.

Unit tests for this commit were also added.