Updated byteorder support in the pcapfile templates and its usage of the osi templates.
Prior to this commit, the pcapfile templates were pretty basic and did
not support a byteorder that was different from the host platform. It
also required the user to explicitly specify the type used to decode the
contents of each packet record.

This commit adds support for decoding all of the integer types in either
byteorder by checking the header with a 0-length type. Support for other
magic numbers that are used to distinguish whether the timestamp is in
microseconds or nanoseconds was also included as part of the byteorder
additions.

Also, the pcapfile templates were modified to explicitly use the osi
templates when decoding packets. This is done by mapping the "linktype"
in the header of the pcap with the osi.datalink enumeration types. A
placeholder type for assisting with composing a pcapfile was also
included in the `pcapfile.Packet` definition.