Added a method to the `pcapnextgen.BlockArray` class from the pcapnextgen templates.
The method that was added, `BlockArray.InterfaceCount`, allows its users
to return the current number of interface descriptors that have been
identified while decoding blocks. The method is implemented by returning
the length of the `BlockArray._interfaces_` class descriptor.

The purpose of this additional method is to allow one to distinguish
when there are no interface descriptors available. This is so that the
`EnhancedPacket` class can determine whether there is enough information
to figure out the data link type that should be used to decode the
packet body. Hence, the "Data" field from the `EnhancedPacket` class was
updated to check the number of interfaces before determining what to do.