Modified the `File` type from the pcapfile templates to honor the provider size.
Prior to this commit, the `File` type would treat the `List` of packets
as an infinite array. This could result in decoding an extra byte prior
to the termination of said array due to how infinite arrays are designed
to work.

This commit modifies the `File` type so that the `List` of packets will
be treated as a block array using the the provider size if it is
available (bounded). If it's not, then we fall back to the original
implementation of using an infinite array.