Fixed an issue with how layers and their sizes are decoded by the osi templates.
Usage of the osi templates are based around the interactions of the
`osi.layers` type and the `osi.stackable` decorator. The `osi.layers`
type determines the size of an individual layer by asking the last
instance of `osi.stackable`. Prior to this commit, the size from the
last stackable layer would always be honored. This is wrong as the size
of any of the previous layers should be honored.

This commit remedies the issue by giving priority to the size returned
by the earlier layers, and then falling back to the size from the
previous layer if none of the earlier layers have mentioned a size.

The `osi.transport.udp` templates were also updated to use the smallest
of either the UDP packet length, or the remaining bytes as specified by
any earlier layers.