Fixed a bug in the `parray.__array_interface__.__len__` method from the `ptypes` library.
The "__len__" method for the `ptypes.parray` class was making a bogus
decision in order to determine whether to return the "length" attribute
of the uninitialized array, or the actual length of the initialized
array. This was due to the implementation of the method using the
`__array_interface__.initializedQ` method for determining initialization
rather than checking whether the "value" attribute is actually defined.
This bug manifests itself only when using an array type that implements
a custom "initializedQ" method, such as the `parray.uninitialized` or
the `parray.block` class.

The aforementioned issue is fixed by explicitly checking if the "value"
attribute is `None` and then returning the "length" attribute if so. If
the "value" is not `None`, then we return the container length as the
length of the array.