Tweaked the allocator for `parray.block` from the ptypes library and added some unit-tests.
The `parray.block` class attempts to describe an array type that has a
length bounded by its "blocksize" attribute. Later in its lifetime, the
class was tweaked to change the way the array is allocated when given an
explicit "length" attribute, or some of values to assign. Unfortunately,
the implementation of this method did not actually work as intended for
all of the available cases.

What's supposed to happen is that if a "length" attribute is anything
other than `None` for a `parray.block`, then that "length" attribute
will be used to allocate the specified number of array elements. If the
"blocksize" method has been defined (or attribute-assigned), the integer
returned by this method will be used to clamp the number of elements
being allocated according to the returned size. If there is no custom
"blocksize" or "length" attribute assigned, then we fall back to
allocating it so that it works like the original `parray.type` class.

This commit corrects the implementation so that it works as as intended,
and applies the aforementioned logic to both `parray.block.alloc` and
the `parray.block.load` methods. In order to distinguish a custom method
being used for the "blocksize", a dummy-implementation of said method
was added to the `parray.block` class which can be checked by using the
protected "__blocksize_originalQ__" method.

To be clear, if there's a custom "blocksize" method attached to the
`parray.block` type and some values are passed to the "alloc" method,
then it will use the super method of the `parray.uninitialized` class
which will continuously load elements until reaching the integer
returned by the "blocksize" method. Otherwise, we instead use the super
method for the `parray.terminated` class. If there isn't a "blocksize"
method implemented, and we were explicitly given arguments or values,
then we go ahead and use the super method of the `parray.terminated`
class. If the instance of `parray.block` has `None` for its length, and
no custom "blocksize" method, then we fall back to using the super
method for the `parray.type` class. This way the instance of the class
will allocate like a regular array, simplifying construction.

This commit also includes 20 unit-tests for exercising a number of cases
and verifying that the implementation works as intended for both
allocating and loading of a `parray.block` instance.