Fixed a bug in the `jfif.Marker` definition from the `image.jpeg` template library.
The `image.jpeg` template library defines an enumeration base class
called `MarkerType`. This class will check its `MarkerType._values_`
attribute to convert it into a list of pairs containing the name and
value for said marker. This is done during the construction of the
`MarkerType` which results in executing `MarkerType.__init__`.

This process could unfortunately result in a situation where some of the
enumeration values do not get converted from their byte form to an
integer that can be used for determining the marker type. This results
in the display for the "Type" field from the `jfif.StreamMarker`
structure being incorrect if there isn't an actual type associated with
the marker.

To remedy the issue, we rewrite the `jfif.MarkerType._values_` list so
that it converts each marker from the specified bytes to an integer at
the time of declaration rather than at the time of construction.