Updated the `image.jpeg.jfif` templates to decode the "APP1" marker.
The application-specific "APP1" marker inside a JPEG file is responsible
for containing metadata associated with an image. This metadata for a
JPEG and is structured using the "Exif" format. Due to a prior commit
introducing support for the `image.jpeg.exif` format, adding support for
decoding this metadata was pretty straightforward.